FEATURE: Background stripes
This is most likely a work in progress. While it works that way, I'm not happy with the empty elements and the linear-gradients make it hard to animate the lines.

Maybe I'll have to add a layer for each of the lines. Forgoing the empty elements is possible but probably not worth the effort.